refactor: create worker with vite inline
